remove middleware.ts, use-mobile.tsx, next.config.js. merged utils.ts and env.ts
