+++ .gitignore
@@ -12,7 +12,7 @@
 /prisma/db.sqlite
 /prisma/db.sqlite-journal
 db.sqlite
-/data/fake-data.ts
+src/data/fake-data.ts
 
 # next.js
 /.next/
