--- next.config.js
@@ -1,31 +1,0 @@
-const config = {
-  reactStrictMode: true,
-  i18n: {
-    locales: ["en"],
-    defaultLocale: "en",
-  },
-  images: {
-    remotePatterns: [
-      {
-        protocol: 'https',
-        hostname: '**.digitaloceanspaces.com',
-        pathname: '/**',
-      },
-      {
-        protocol: 'http',
-        hostname: 'localhost',
-        port: '9000',
-        pathname: '/**',
-      },
-      {
-        protocol: 'http',
-        hostname: '127.0.0.1',
-        port: '9000',
-        pathname: '/**',
-      },
-    ],
-  },
-  transpilePackages: ['next-auth'],
-};
-
-export default config;
